requirements.txt
fastapi
gradio
google-generativeai
SpeechRecognition
pyttsx3
uvicorn
